fix: use Int64 for var_ptr to prevent overflow on large proofs
Changed PBSystem.var_ptr from Vector{Int32} to Vector{Int64}.
Like row_ptr, var_ptr stores cumulative positions - in this case,
positions into the var_eqs array for the reverse index (variable → equations).

On large proofs with millions of variables each appearing in hundreds
of equations, the total number of (variable, equation) pairs can exceed
2^31-1. The earlier row_ptr overflow at 2.1B literals proves that
var_ptr will encounter the same issue since var_eqs has similar size.

+++ trimnalyser.jl
@@ -992,7 +992,7 @@ end; # using .Dumping # to save the import un comment this.
         rhs     ::Vector{Int64}
         row_ptr ::Vector{Int64}
         # Inverse: variable → equations containing it
-        var_ptr ::Vector{Int32}     # length = n_vars + 1
+        var_ptr ::Vector{Int64}     # length = n_vars + 1
         var_eqs ::Vector{Int32} end # flat list of equation ids
 
     mutable struct Trail
@@ -1125,7 +1125,7 @@ end; # using .Dumping # to save the import un comment this.
 
         var_count = zeros(Int32, n_vars)
         for v in vars; var_count[v] += 1; end
-        var_ptr = Vector{Int32}(undef, n_vars + 1)
+        var_ptr = Vector{Int64}(undef, n_vars + 1)
         var_ptr[1] = 1
         for v in 1:n_vars
             var_ptr[v+1] = var_ptr[v] + var_count[v]
